+++ crates/common/src/lib.rs
@@ -346,6 +346,7 @@ impl From<BlockHash> for BlockId {
 pub enum EthereumChain {
     Mainnet,
     Goerli,
+    Sepolia,
     Other(primitive_types::U256),
 }
 
@@ -353,8 +354,10 @@ pub enum EthereumChain {
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum Chain {
     Mainnet,
-    Testnet,
-    Integration,
+    GoerliTestnet,
+    GoerliIntegration,
+    SepoliaTestnet,
+    SepoliaIntegration,
     Custom,
 }
 
@@ -375,16 +378,20 @@ impl ChainId {
     }
 
     pub const MAINNET: Self = Self::from_slice_unwrap(b"SN_MAIN");
-    pub const TESTNET: Self = Self::from_slice_unwrap(b"SN_GOERLI");
-    pub const INTEGRATION: Self = Self::from_slice_unwrap(b"SN_GOERLI");
+    pub const GOERLI_TESTNET: Self = Self::from_slice_unwrap(b"SN_GOERLI");
+    pub const GOERLI_INTEGRATION: Self = Self::from_slice_unwrap(b"SN_GOERLI");
+    pub const SEPOLIA_TESTNET: Self = Self::from_slice_unwrap(b"SN_SEPOLIA");
+    pub const SEPOLIA_INTEGRATION: Self = Self::from_slice_unwrap(b"SN_INTEGRATION_SEPOLIA");
 }
 
 impl std::fmt::Display for Chain {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
             Chain::Mainnet => f.write_str("Mainnet"),
-            Chain::Testnet => f.write_str("Görli"),
-            Chain::Integration => f.write_str("Integration"),
+            Chain::GoerliTestnet => f.write_str("Testnet/Görli"),
+            Chain::GoerliIntegration => f.write_str("Integration/Görli"),
+            Chain::SepoliaTestnet => f.write_str("Testnet/Sepolia"),
+            Chain::SepoliaIntegration => f.write_str("Integration/Sepolia"),
             Chain::Custom => f.write_str("Custom"),
         }
     }