+++ crates/gateway-client/src/lib.rs
@@ -329,14 +329,25 @@ impl Client {
         Self::with_base_url(Url::parse("https://alpha-mainnet.starknet.io/").unwrap()).unwrap()
     }
 
-    /// Creates a [Client] for [Chain::Testnet].
-    pub fn testnet() -> Self {
+    /// Creates a [Client] for [Chain::GoerliTestnet].
+    pub fn goerli_testnet() -> Self {
         Self::with_base_url(Url::parse("https://alpha4.starknet.io/").unwrap()).unwrap()
     }
 
-    /// Creates a [Client] for [Chain::Integration].
-    pub fn integration() -> Self {
+    /// Creates a [Client] for [Chain::GoerliIntegration].
+    pub fn goerli_integration() -> Self {
         Self::with_base_url(Url::parse("https://external.integration.starknet.io").unwrap())
+            .unwrap()
+    }
+
+    /// Creates a [Client] for [Chain::SepoliaTestnet].
+    pub fn sepolia_testnet() -> Self {
+        Self::with_base_url(Url::parse("https://alpha-sepolia.starknet.io/").unwrap()).unwrap()
+    }
+
+    /// Creates a [Client] for [Chain::SepoliaIntegration].
+    pub fn sepolia_integration() -> Self {
+        Self::with_base_url(Url::parse("https://integration-sepolia.starknet.io/").unwrap())
             .unwrap()
     }
 
@@ -394,7 +405,7 @@ impl Client {
     /// Returns the [network chain](Chain) this client is operating on.
     pub async fn chain(&self) -> anyhow::Result<Chain> {
         use pathfinder_common::consts::{
-            INTEGRATION_GENESIS_HASH, MAINNET_GENESIS_HASH, TESTNET_GENESIS_HASH,
+            GOERLI_INTEGRATION_GENESIS_HASH, GOERLI_TESTNET_GENESIS_HASH, MAINNET_GENESIS_HASH,
         };
         // unwrap is safe as `block_hash` is always present for non-pending blocks.
         let genesis_hash = self
@@ -405,9 +416,11 @@ impl Client {
             .block_hash;
 
         match genesis_hash {
-            testnet if testnet == TESTNET_GENESIS_HASH => Ok(Chain::Testnet),
+            testnet if testnet == GOERLI_TESTNET_GENESIS_HASH => Ok(Chain::GoerliTestnet),
             mainnet if mainnet == MAINNET_GENESIS_HASH => Ok(Chain::Mainnet),
-            integration if integration == INTEGRATION_GENESIS_HASH => Ok(Chain::Integration),
+            integration if integration == GOERLI_INTEGRATION_GENESIS_HASH => {
+                Ok(Chain::GoerliIntegration)
+            }
             other => Err(anyhow::anyhow!("Unknown genesis block hash: {}", other.0)),
         }
     }
@@ -731,9 +744,9 @@ pub mod test_utils {
         S2: std::string::ToString + Send + Sync + Clone + 'static,
     {
         if std::env::var_os("SEQUENCER_TESTS_LIVE_API").is_some() {
-            (None, Client::testnet())
+            (None, Client::goerli_testnet())
         } else if std::env::var_os("SEQUENCER_TESTS_LIVE_API_INTEGRATION").is_some() {
-            (None, Client::integration())
+            (None, Client::goerli_integration())
         } else {
             use warp::Filter;
             let opt_query_raw = warp::query::raw()
@@ -1707,7 +1720,9 @@ mod tests {
             {
                 match target {
                     TargetChain::Mainnet => (None, Client::mainnet().disable_retry_for_tests()),
-                    TargetChain::Testnet => (None, Client::testnet().disable_retry_for_tests()),
+                    TargetChain::Testnet => {
+                        (None, Client::goerli_testnet().disable_retry_for_tests())
+                    }
                     // Escaped above already
                     TargetChain::Invalid => unreachable!(),
                 }
@@ -1763,7 +1778,7 @@ mod tests {
         async fn testnet() {
             let (_server_handle, sequencer) = setup_server(TargetChain::Testnet);
             let chain = sequencer.chain().await.unwrap();
-            assert_eq!(chain, Chain::Testnet);
+            assert_eq!(chain, Chain::GoerliTestnet);
         }
 
         #[tokio::test]