+++ CHANGELOG.md
@@ -15,10 +15,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Configuration options to selectively enable/disable parts of the node. This can be useful to run tests or benchmarks with isolated components e.g. test RPC methods without the sync process updating the database.
   - `rpc.enable` configuration option to enable/disable the RPC server. Defaults to enabled.
   - `sync.enable` configuration option to enable/disable the sync process. Defaults to enabled. 
+- Support for Sepolia testnet via `--network testnet-sepolia`
+- Support for Sepolia integration via `--network integration-sepolia`
 
 ### Changed
 
 - Default RPC version is now v0.5 (was v0.4). This can be manually configured to any version.
+- Goerli testnet network selection is now `--network testnet-goerli`, removed `--network testnet`
+- Goerli integration network selection is now `--network integration-goerli`, removed `--network testnet`
 
 ### Removed